One max-file-size command and script to rule them all
